// SIC/SIC Admin/src/types/seasonalTopics.ts
export type SeasonalTopicDataArr = {
	id: number;
	uuid: string;
	name: string;
	namePinging: string;
	nameSimplifiedChinese: string;
	nameTraditionalChinese: string;
	image: string;
	rewardStars: number;
	startDate: string;
	endDate: string;
	order: number;
	isActive: boolean;
	createdAt: string;
	updatedAt: string;
	createdBy: number;
};

export type AddEditSeasonalTopicData = {
	onClose: () => void;
	editData: SeasonalTopicDataArr | null;
	disableData: boolean;
	onSubmit: () => void;
};

export type CreateSeasonalTopic = {
	name: string;
	namePinging: string;
	nameTraditionalChinese: string;
	nameSimplifiedChinese: string;
	rewardStars: number | string;
	image: File | null;
	startDate: string;
	endDate: string;
	isForSeasonal: boolean;
};